- 코드 정리

+++ src/main/java/com/mindplates/everyonesprint/framework/websocket/WebSocketEventListener.java
@@ -10,8 +10,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.stereotype.Component;
-import org.springframework.web.socket.messaging.SessionConnectedEvent;
 import org.springframework.web.socket.messaging.SessionDisconnectEvent;
+
+import java.util.Map;
 
 @Component
 @Slf4j
@@ -30,20 +31,18 @@ public class WebSocketEventListener {
     }
 
     @EventListener
-    public void handleWebSocketConnectListener(SessionConnectedEvent event) {
-    }
-
-    @EventListener
     public void handleWebSocketDisconnectListener(SessionDisconnectEvent event) {
         StompHeaderAccessor headerAccessor = StompHeaderAccessor.wrap(event.getMessage());
-        UserSession userSession = (UserSession) headerAccessor.getSessionAttributes().get("USER_INFO");
-        if (userSession != null) {
-            MessageData data = MessageData.builder().type("PUBLIC-PARK-EXIT").build();
-            messageSendService.sendTo("public-park", data, userSession);
-            walkerService.deleteById(userSession.getId().toString());
+        Map<String, Object> attributes = headerAccessor.getSessionAttributes();
+        if (attributes != null) {
+            UserSession userSession = (UserSession) attributes.get("USER_INFO");
+            if (userSession != null) {
+                MessageData data = MessageData.builder().type("PUBLIC-PARK-EXIT").build();
+                messageSendService.sendTo("public-park", data, userSession);
+                walkerService.deleteById(userSession.getId().toString());
 
-            meetingService.updateUserLeaveInfo(headerAccessor.getSessionId(), userSession);
+                meetingService.updateUserLeaveInfo(headerAccessor.getSessionId(), userSession);
+            }
         }
-
     }
 }
